Add samples for all platforms
