--- richeditor/src/commonMain/kotlin/com/mocoding/richeditor/model/RichTextPart.kt
@@ -1,7 +1,0 @@
-package com.mocoding.richeditor.model
-
-data class RichTextPart(
-    val fromIndex: Int,
-    val toIndex: Int,
-    val styles: Set<RichTextStyle>,
-)
