+++ richeditor/src/commonMain/kotlin/com/mohamedrejeb/richeditor/model/RichTextPart.kt
@@ -1,0 +1,7 @@
+package com.mohamedrejeb.richeditor.model
+
+data class RichTextPart(
+    val fromIndex: Int,
+    val toIndex: Int,
+    val styles: Set<RichTextStyle>,
+)
