+++ sample/android/src/main/AndroidManifest.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.mocoding.richeditor.android">
+<manifest xmlns:android="http://schemas.android.com/apk/res/android" package="com.mohamedrejeb.richeditor.android">
     <application
             android:allowBackup="false"
             android:supportsRtl="true"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
-        <activity android:name="com.mocoding.richeditor.sample.MainActivity" android:exported="true">
+        <activity android:name="com.mohamedrejeb.richeditor.sample.MainActivity" android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
                 <category android:name="android.intent.category.LAUNCHER"/>
