+++ sample/android/src/main/java/com/mohamedrejeb/richeditor/sample/MainActivity.kt
@@ -1,0 +1,18 @@
+package com.mohamedrejeb.richeditor.sample
+
+import android.os.Bundle
+import androidx.activity.compose.setContent
+import androidx.appcompat.app.AppCompatActivity
+import androidx.compose.material.MaterialTheme
+import com.mohamedrejeb.richeditor.sample.common.App
+
+class MainActivity : AppCompatActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            MaterialTheme {
+                App()
+            }
+        }
+    }
+}
