--- sample/common/src/commonMain/kotlin/com/mocoding/richeditor/sample/common/ui.theme/Typography.kt
@@ -1,16 +1,0 @@
-package com.mocoding.richeditor.sample.common.ui.theme
-
-//val Typography = Typography(
-    /* Other default text styles to override
-    button = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.W500,
-        fontSize = 14.sp
-    ),
-    caption = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 12.sp
-    )
-    */
-//)
