+++ sample/common/src/commonMain/kotlin/com/mohamedrejeb/richeditor/sample/common/CustomStyle.kt
@@ -1,0 +1,4 @@
+package com.mohamedrejeb.richeditor.sample.common
+
+class CustomStyle {
+}
