+++ sample/common/src/commonMain/kotlin/com/mohamedrejeb/richeditor/sample/common/RichTextStyleButton.kt
@@ -1,0 +1,54 @@
+package com.mohamedrejeb.richeditor.sample.common
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.focus.focusProperties
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import com.mohamedrejeb.richeditor.model.RichTextStyle
+import com.mohamedrejeb.richeditor.model.RichTextValue
+
+@Composable
+fun RichTextStyleButton(
+    style: RichTextStyle,
+    value: RichTextValue,
+    onValueChanged: (RichTextValue) -> Unit,
+    icon: ImageVector,
+    tint: Color? = null
+) {
+    IconButton(
+        modifier = Modifier
+            // Workaround to prevent the rich editor
+            // from losing focus when clicking on the button
+            // (Happens only on Desktop)
+            .focusProperties { canFocus = false },
+        onClick = {
+            onValueChanged(value.toggleStyle(style))
+        },
+        colors = IconButtonDefaults.iconButtonColors(
+            contentColor = if (value.currentStyles.contains(style)) {
+                MaterialTheme.colorScheme.onPrimary
+            } else {
+                MaterialTheme.colorScheme.onBackground
+            },
+        ),
+    ) {
+        Icon(
+            icon,
+            contentDescription = icon.name,
+            tint = tint ?: LocalContentColor.current,
+            modifier = Modifier
+                .background(
+                    color = if (value.currentStyles.contains(style)) {
+                        MaterialTheme.colorScheme.primary
+                    } else {
+                        Color.Transparent
+                    },
+                    shape = CircleShape
+                )
+        )
+    }
+}
