+++ sample/common/src/commonMain/kotlin/com/mohamedrejeb/richeditor/sample/common/ui.theme/Color.kt
@@ -1,0 +1,46 @@
+package com.mohamedrejeb.richeditor.sample.common.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Black = Color(0xFF090f0b)
+
+val Green300 = Color(0xFF2EB688)
+val Green400 = Color(0xFF145526)
+val Green500 = Color(0xFF046D4A)
+
+val Red300 = Color(0xFFF33736)
+val Red400 = Color(0xFFcb290b)
+val Red500 = Color(0xFF9C2221)
+
+val Blue300 = Color(0xFF54B1DF)
+val Blue500 = Color(0xFF1E3DA8)
+
+val Yellow300 = Color(0xFFF1A22C)
+val Yellow400 = Color(0xFFfaae41)
+val Yellow500 = Color(0xFFCB5C0D)
+
+val Blue400 = Color(0xFF4572E8)
+
+val LightGray400 = Color(0xFFb8b6b3)
+
+val DarkGray400 = Color(0xFF3e4047)
+
+val Gray400 = Color(0xFF595C61)
+
+val Bug = Color(0xFF179A55)
+val Dark = Color(0xFF040706)
+val Dragon = Color(0xFF378A94)
+val Electric = Color(0xFFE0E64B)
+val Fairy = Color(0xFF9E1A44)
+val Fire = Color(0xFFB22328)
+val Flying = Color(0xFF90B1C5)
+val Ghost = Color(0xFF363069)
+val Ice = Color(0xFF7ECFF2)
+val Poison = Color(0xFF642785)
+val Psychic = Color(0xFFAC296B)
+val Rock = Color(0xFF4B190E)
+val Steel = Color(0xFF5C756D)
+val Water = Color(0xFF2648DC)
+val Fighting = Color(0xFF9F422A)
+val Grass = Color(0xFF007C42)
+val Ground = Color(0xFFAD7235)
