+++ sample/common/src/commonMain/kotlin/com/mohamedrejeb/richeditor/sample/common/ui.theme/Typography.kt
@@ -1,0 +1,18 @@
+package com.mohamedrejeb.richeditor.sample.common.ui.theme
+
+import androidx.compose.material3.Typography
+
+val Typography = Typography(
+    /* Other default text styles to override
+    button = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.W500,
+        fontSize = 14.sp
+    ),
+    caption = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 12.sp
+    )
+    */
+)
