+++ sample/web/src/jsMain/kotlin/Main.kt
@@ -2,7 +2,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.window.Window
-import com.mocoding.richeditor.sample.common.App
+import com.mohamedrejeb.richeditor.sample.common.App
 import org.jetbrains.skiko.wasm.onWasmReady
 
 fun main() {
